db/requirements.txt
# db/requirements.txt — Python dependencies for migrations and seeder
psycopg2-binary>=2.9
python-dotenv>=1.0
